chore: subscriber as separate pacakge
